Move visited plugin hook handler code
